backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.7
django-filter==25.2
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
django-cors-headers>=4.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9